Flatten the hierarchy
