+++ lib/foreman_tasks_core/task_launcher/batch.rb
@@ -1,24 +1,5 @@
 module ForemanTasksCore
   module TaskLauncher
-    class ParentAction < ::Dynflow::Action
-      include Dynflow::Action::WithSubPlans
-      include Dynflow::Action::WithPollingSubPlans
-
-      # { task_id => { :action_class => Klass, :input => input } }
-      def plan(launcher, input_hash)
-        launcher.launch_children(self, input_hash)
-        plan_self
-      end
-
-      def initiate
-        ping suspended_action
-        wait_for_sub_plans sub_plans
-      end
-
-      def rescue_strategy
-        Dynflow::Action::Rescue::Fail
-      end
-    end
 
     class Batch < Abstract
       def launch!(input)
